feat: Add new goal tree to profile and subgoals pages
