--- assets/js/features/goals/GoalTree/tree-v2.tsx
@@ -1,122 +1,0 @@
-import React, { useState } from "react";
-
-import { IconChevronDown, IconChevronRight } from "@tabler/icons-react";
-import { createTestId } from "@/utils/testid";
-
-import { useTreeContext, TreeContextProvider, TreeContextProviderProps } from "./treeContext";
-import { ExpandGoalSuccessConditions, GoalActions, GoalDetails, GoalProgressBar } from "./components/GoalDetails";
-import { ProjectDetails } from "./components/ProjectDetails";
-import { GoalNode, Node, ProjectNode } from "./tree";
-import { useExpandable } from "./context/Expandable";
-import { NodeIcon } from "./components/NodeIcon";
-import { NodeName } from "./components/NodeName";
-import { Controls } from "./components/Controls-v2";
-
-export function GoalTree(props: TreeContextProviderProps) {
-  return (
-    <TreeContextProvider {...props}>
-      <GoalTreeRoots />
-    </TreeContextProvider>
-  );
-}
-
-function GoalTreeRoots() {
-  const context = useTreeContext();
-
-  return (
-    <div>
-      <Controls />
-
-      <div className="border-b border-surface-outline">
-        {context.tree.map((root) => (
-          <NodeView key={root.id} node={root} />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function NodeView({ node }: { node: Node }) {
-  return (
-    <div>
-      {node.type === "goal" ? <GoalHeader node={node as GoalNode} /> : <ProjectHeader node={node as ProjectNode} />}
-      <NodeChildren node={node} />
-    </div>
-  );
-}
-
-function ProjectHeader({ node }: { node: ProjectNode }) {
-  const testId = createTestId("project", node.project.name!);
-
-  return (
-    <HeaderContainer node={node} data-test-id={testId}>
-      <div className="flex items-center gap-1">
-        <NodeIcon node={node} />
-        <NodeName node={node} />
-      </div>
-
-      <ProjectDetails node={node} />
-    </HeaderContainer>
-  );
-}
-
-function GoalHeader({ node }: { node: GoalNode }) {
-  const [hovered, setHovered] = useState(false);
-  const testId = createTestId("goal", node.goal.name!);
-
-  return (
-    <HeaderContainer
-      node={node}
-      onMouseEnter={() => setHovered(true)}
-      onMouseLeave={() => setHovered(false)}
-      data-test-id={testId}
-    >
-      <div className="flex items-center gap-1 relative">
-        <NodeExpandCollapseToggle node={node} />
-        <NodeIcon node={node} />
-        <NodeName node={node} />
-        <GoalProgressBar node={node} />
-        <ExpandGoalSuccessConditions node={node} />
-        <GoalActions node={node} hovered={hovered} />
-      </div>
-
-      <GoalDetails node={node} />
-    </HeaderContainer>
-  );
-}
-
-function NodeChildren({ node }: { node: Node }) {
-  const { expanded } = useExpandable();
-
-  if (!expanded[node.id] || !node.hasChildren) return <></>;
-
-  return <>{node.children?.map((node) => <NodeView key={node.id} node={node} />)}</>;
-}
-
-function NodeExpandCollapseToggle({ node }: { node: Node }) {
-  const { expanded, toggleExpanded } = useExpandable();
-
-  if (node.children.length === 0) return <></>;
-
-  const resourceId = "goal" in node ? (node as GoalNode).goal.id : (node as ProjectNode).project.id;
-  const testId = createTestId("toggle-node", resourceId!);
-  const handleClick = () => toggleExpanded(node.id);
-  const ChevronIcon = expanded[node.id] ? IconChevronDown : IconChevronRight;
-
-  return (
-    <ChevronIcon
-      size={16}
-      className="absolute left-[-1.2rem] top-1 cursor-pointer"
-      onClick={handleClick}
-      data-test-id={testId}
-    />
-  );
-}
-
-function HeaderContainer(props) {
-  return (
-    <div className="border-t border-surface-outline py-3" {...props}>
-      <div style={{ paddingLeft: props.node.depth * 30 }}>{props.children}</div>
-    </div>
-  );
-}
